+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 0.0.45
 
+* `Add support of containers in `UIKitContainer` modifiers and `Container` function
+
 ## 0.0.44
 
 * `Versions`: